Finally working client and server
Futures are meant to be polled multiple times, so take the result
only once it's ready. BufReader was polling the AsyncRead Stream
twice and then the completion event wasn't firing properly.

A bunch of tracing, too much tracing. I need to pare this back somehow.

Next there is a bug with borrowing the rcrefcell map of subscribers
across an await point.