Implemented file
OpenAI codex implemented src/file.rs and the example, but I had to
polish it up as file_example didn't work. This is because it was using
read_to_end and the way we do File and TcpStream without internal
buffers creates problems.

Basically because a buffer passed into poll_read may change when it
returns Poll::Pending but the pointer has already been passed to the
kernel through io uring means that it's not really valid. To get around
this we'd want to internally buffer first.